[+] structured root helpers, added random_range function, created shuffles section -> move it to specific branch
