+++ src/shuffles/fisher_yetes_shuffle/fisher_yetes_shuffle.ts
@@ -1,0 +1,9 @@
+function fisher_yetes_shuffle(data_set: Array<any>): Array<any> {
+  if (data_set.length <= 1)
+    throw new Error("For shuffling need to have more than 1 element");
+
+  for (let i = 0; i < data_set.length; i++) {}
+  return [];
+}
+
+export default fisher_yetes_shuffle;
